refactor: rename arch tests to have the `ArchTest` suffix
