+++ qa/archunit-tests/src/test/java/io/camunda/client/DefaultTenantAwarenessArchTest.java
@@ -1,17 +1,9 @@
 /*
- * Copyright © 2017 camunda services GmbH (info@camunda.com)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
+ * one or more contributor license agreements. See the NOTICE file distributed
+ * with this work for additional information regarding copyright ownership.
+ * Licensed under the Camunda License 1.0. You may not use this file
+ * except in compliance with the Camunda License 1.0.
  */
 package io.camunda.client;
 
@@ -33,7 +25,7 @@ import java.util.Set;
 @AnalyzeClasses(
     packages = "io.camunda.client",
     importOptions = ImportOption.DoNotIncludeTests.class)
-public class DefaultTenantAwarenessTest {
+public class DefaultTenantAwarenessArchTest {
 
   /**
    * Tries to ensure that every command implementation that is tenant-aware respects the configured
@@ -70,7 +62,7 @@ public class DefaultTenantAwarenessTest {
 
         final boolean callsGetDefault =
             allCallsFromConstructor.stream()
-                .anyMatch(DefaultTenantAwarenessTest::isGetDefaultTenantIdCall);
+                .anyMatch(DefaultTenantAwarenessArchTest::isGetDefaultTenantIdCall);
         if (!callsGetDefault) {
           events.add(
               SimpleConditionEvent.violated(
