+++ qa/archunit-tests/src/test/java/io/camunda/zeebe/gateway/rest/ControllerStrictJsonCompareArchTest.java
@@ -5,26 +5,24 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.client;
-
-import static com.tngtech.archunit.lang.syntax.ArchRuleDefinition.noClasses;
+package io.camunda.zeebe.gateway.rest;
 
 import com.tngtech.archunit.core.importer.ImportOption;
 import com.tngtech.archunit.junit.AnalyzeClasses;
 import com.tngtech.archunit.junit.ArchTest;
 import com.tngtech.archunit.lang.ArchRule;
+import com.tngtech.archunit.lang.syntax.ArchRuleDefinition;
+import org.springframework.test.web.reactive.server.WebTestClient.BodyContentSpec;
 
 @AnalyzeClasses(
-    packages = "io.camunda.client",
-    importOptions = ImportOption.DoNotIncludeTests.class)
-public class ArchitectureTest {
+    packages = "io.camunda.zeebe.gateway.rest",
+    importOptions = ImportOption.OnlyIncludeTests.class)
+public class ControllerStrictJsonCompareArchTest {
 
+  /** This ArchUnit test ensures that any REST API controller tests use JsonCompareMode.STRICT */
   @ArchTest
-  public static final ArchRule RULE_CLIENT_API_CLASSES_MUST_NOT_DEPEND_ON_PROTOCOL_REST_PACKAGE =
-      noClasses()
-          .that()
-          .resideInAPackage("io.camunda.client.api..")
+  public static final ArchRule RULE_USE_STRICT_JSON_COMPARISON =
+      ArchRuleDefinition.noClasses()
           .should()
-          .dependOnClassesThat()
-          .resideInAPackage("io.camunda.client.protocol.rest..");
+          .callMethod(BodyContentSpec.class, "json", String.class);
 }
